src/postgres/requirements.txt
psycopg2-binary==2.9.9
pandas==2.2.2
pymongo==4.6.3